+++ src/clients/catalog.ts
@@ -1,13 +1,18 @@
-import {
+import type {
   InstanceOptions,
   IOContext,
-  JanusClient,
   RequestTracingConfig,
 } from '@vtex/api'
+import { JanusClient } from '@vtex/api'
 
 import { checkSellerInformation } from '../utils/seller'
-import { AuthMethod } from '../typings/tokens'
-import { GetSkuResponse, Seller } from '../typings/catalog'
+import type { AuthMethod } from '../typings/tokens'
+import type {
+  GetSkuResponse,
+  GetSkuContextResponse,
+  Seller,
+  Category,
+} from '../typings/catalog'
 import { getRequestConfig } from '../utils/request'
 
 const baseURL = '/api/catalog'
@@ -15,7 +20,11 @@ const baseURLLegacy = '/api/catalog_system'
 
 const routes = {
   productsAndSkus: `${baseURLLegacy}/pvt/products/GetProductAndSkuIds`,
+  getCategoryById: (categoryId: string) =>
+    `${baseURL}/pvt/category/${categoryId}`,
   getSkuById: (skuId: string) => `${baseURL}/pvt/stockkeepingunit/${skuId}`,
+  getSkuContext: (skuId: string) =>
+    `${baseURLLegacy}/pvt/sku/stockkeepingunitbyid/${skuId}`,
   changeNotification: (sellerId: string, skuId: string) =>
     `${baseURLLegacy}/pvt/skuseller/changenotification/${sellerId}/${skuId}`,
   seller: (sellerId: string) => `${baseURLLegacy}/pvt/seller/${sellerId}`,
@@ -27,6 +36,19 @@ export class Catalog extends JanusClient {
     super(ctx, {
       ...options,
     })
+  }
+
+  public getCategoryById(
+    categoryId: string,
+    authMethod: AuthMethod = 'AUTH_TOKEN',
+    tracingConfig?: RequestTracingConfig
+  ): Promise<Category | undefined> {
+    const metric = 'catalog-getCategoryMetric'
+
+    return this.http.get<Category>(
+      routes.getCategoryById(categoryId),
+      getRequestConfig(this.context, authMethod, metric, tracingConfig)
+    )
   }
 
   public getProductsAndSkus(
@@ -50,6 +72,19 @@ export class Catalog extends JanusClient {
 
     return this.http.get<GetSkuResponse>(
       routes.getSkuById(skuId),
+      getRequestConfig(this.context, authMethod, metric, tracingConfig)
+    )
+  }
+
+  public getSkuContext(
+    skuId: string,
+    authMethod: AuthMethod = 'AUTH_TOKEN',
+    tracingConfig?: RequestTracingConfig
+  ) {
+    const metric = 'catalog-getSkuContextMetric'
+
+    return this.http.get<GetSkuContextResponse>(
+      routes.getSkuContext(skuId),
       getRequestConfig(this.context, authMethod, metric, tracingConfig)
     )
   }